category filter at explore working. Category filter can be applied on top of search result.
