+++ client/store/events.js
@@ -45,6 +45,7 @@ export const searchKeyword =
   ({ name, location, date }) =>
   async (dispatch) => {
     try {
+      console.log(name, location, date);
       const events = (await axios.get(`/api/events/search?keyword=${name}&location=${location}&date=${date}`)).data;
       if (events.length === 0) {
         window.localStorage.setItem("error", JSON.stringify({ output: "Result not found." }));
